+++ CHANGELOG.md
@@ -2,12 +2,20 @@
 
 ## [Unreleased]
 
+## [v1.1.0] - 2023-01-14
+
+### Added
+
+- Lint checks that ensure `NonCallableMock` and `AsyncMock` constructors have
+  the `spec` or `spec_set` argument
+
 ## [v1.0.0] - 2023-01-14
 
 ### Added
 
 - Lint checks that ensure `Mock` and `MagicMock` constructors have the `spec`
-  argument
+  or `spec_set`argument
 
 [//]: # "Release links"
 [v1.0.0]: https://github.com/jdkandersson/flake8-mock-spec/releases/v1.0.0
+[v1.1.0]: https://github.com/jdkandersson/flake8-mock-spec/releases/v1.1.0